test/src/requirements.txt
# python Python 3.11.13 (recommended) || 3.10
fastapi
uvicorn
python-dotenv
python-keycloak
PyJWT
alembic 
sqlalchemy 
psycopg2-binary
pytest
tenacity
status

# alembic==1.16.5
# annotated-types==0.7.0
# anyio==4.10.0
# certifi==2025.8.3
# cffi==2.0.0
# charset-normalizer==3.4.3
# click==8.2.1
# cryptography==46.0.1
# dotenv==0.9.9
# fastapi==0.116.2
# greenlet==3.2.4
# h11==0.16.0
# idna==3.10
# iniconfig==2.1.0
# jwt==1.4.0
# Mako==1.3.10
# MarkupSafe==3.0.3
# packaging==25.0
# pluggy==1.6.0
# psycopg2==2.9.10
# psycopg2-binary==2.9.10
# pycparser==2.23
# pydantic==2.11.9
# pydantic_core==2.33.2
# Pygments==2.19.2
# PyJWT==2.10.1
# pytest==8.4.2
# python-dotenv==1.1.1
# requests==2.32.5
# sniffio==1.3.1
# SQLAlchemy==2.0.43
# starlette==0.48.0
# typing-inspection==0.4.1
# typing_extensions==4.15.0
# urllib3==2.5.0
# uvicorn==0.35.0
